Reduce Page object parameter passing to field parameter passing
