+++ src/document_chunker/strategies/line_based.py
@@ -1,39 +1,18 @@
-from abc import ABC, abstractmethod
 from typing import List
 
 from textractor.entities.bbox import BoundingBox
 
 from data_models.chunk_models import DocumentMetadata, OpenSearchChunk
+from document_chunker.strategies.base import ChunkingStrategyHandler
 from document_chunker.utils.bbox_utils import combine_bounding_boxes
-
-
-class ChunkingStrategyHandler(ABC):
-    """Abstract base class for chunking strategies."""
-
-    def __init__(self, maximum_chunk_size: int):
-        self.maximum_chunk_size = maximum_chunk_size
-
-    @abstractmethod
-    def chunk(self, layout_block, page, metadata: DocumentMetadata, chunk_index_start: int) -> List[OpenSearchChunk]:
-        """
-        Extracts chunks from a single layout block based on the specific strategy.
-
-        Args:
-            layout_block: The Textractor LayoutBlock to process.
-            page: The parent Page of the layout_block.
-            metadata: Document metadata.
-            chunk_index_start: The starting index for the chunks produced by this block.
-
-        Returns:
-            A list of OpenSearchChunk objects.
-        """
-        pass
 
 
 class LineBasedChunkingHandler(ChunkingStrategyHandler):
     """Implements the line-based chunking strategy."""
 
-    def chunk(self, layout_block, page, metadata: DocumentMetadata, chunk_index_start: int) -> List[OpenSearchChunk]:
+    def chunk(
+        self, layout_block, page_number: int, metadata: DocumentMetadata, chunk_index_start: int
+    ) -> List[OpenSearchChunk]:
         """Extract chunks using line-based splitting strategy."""
         chunks = []
         chunk_index = chunk_index_start
@@ -48,7 +27,7 @@ class LineBasedChunkingHandler(ChunkingStrategyHandler):
             if self._would_exceed_size_limit(current_chunk_lines, line_text):
                 if current_chunk_lines:
                     chunk = self._create_chunk_from_lines(
-                        current_chunk_lines, current_chunk_bboxes, layout_block, page, metadata, chunk_index
+                        current_chunk_lines, current_chunk_bboxes, layout_block, page_number, metadata, chunk_index
                     )
                     chunks.append(chunk)
                     chunk_index += 1
@@ -60,7 +39,7 @@ class LineBasedChunkingHandler(ChunkingStrategyHandler):
 
         if current_chunk_lines:
             chunk = self._create_chunk_from_lines(
-                current_chunk_lines, current_chunk_bboxes, layout_block, page, metadata, chunk_index
+                current_chunk_lines, current_chunk_bboxes, layout_block, page_number, metadata, chunk_index
             )
             chunks.append(chunk)
 
@@ -78,7 +57,7 @@ class LineBasedChunkingHandler(ChunkingStrategyHandler):
         lines: List[str],
         bboxes: List[BoundingBox],
         layout_block,
-        page,
+        page_number,
         metadata: DocumentMetadata,
         chunk_index: int,
     ) -> OpenSearchChunk:
@@ -88,7 +67,7 @@ class LineBasedChunkingHandler(ChunkingStrategyHandler):
 
         return OpenSearchChunk.from_textractor_layout_and_text(
             block=layout_block,
-            page=page,
+            page_num=page_number,
             metadata=metadata,
             chunk_index=chunk_index,
             chunk_text=chunk_text,
